Wordcloud/requirements.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.0
pandas==1.2.1
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
urllib3==1.26.3
wordcloud==1.8.1
